requirements.txt
pip>=19.2.3
bump2version>=0.5.11
wheel>=0.33.6
watchdog>=0.9.0
flake8>=3.7.8
tox>=3.14.0
coverage>=4.5.4
Sphinx>=1.8.5
sphinx-rtd-theme>=1.0.0
twine>=1.14.0
mypy>=0.910
black>=21.7b0
pytest>=6.2.4
scipy>=1.5.0
numpy>=1.13.0
typing-extensions>=3.10.0
pycairo>=1.20.0
opencv-python>=4.5.0
numpy-hilbert-curve>=1.0.0
descartes>=1.1.0
alphashape>=1.3.1
scikit-learn>=0.24.0
tqdm>=4.60.0